Added development module, proper tests, fixed issues in lib and schema
tests draw the schema from tests/schema.json for now. Final bit of serialisation (i.e. writing the json string to file) is still pending and needs {mzqc: ...} wrap.
