+++ pylib/tests/schema.json
@@ -1,0 +1,240 @@
+{
+	"$schema": "http://json-schema.org/draft-07/schema#",
+	"$id": "https://github.com/HUPO-PSI/qcML-development/tree/master/schema/v0_0_11/mzqc_0_0_11.schema.json",
+	"title": "mzQC schema v0.0.11",
+	"description": "JSON schema specifying the mzQC format v0.0.11 developed by the HUPO-PSI Quality Control working group (http://psidev.info/groups/quality-control).",
+	"type": "object",
+	"properties": {
+		"mzQC": {
+			"description": "Root mzQC element.",
+			"type": "object",
+			"properties": {
+				"version": {
+					"description": "Version number of the relevant mzQC JSON schema.",
+					"type": "string",
+					"pattern": "^\\d+\\.\\d+\\.\\d+$"
+				},
+				"creationDate": {
+					"description": "The date and time at which the mzQC file was created (date-time in isoformat).",
+					"type": "string",
+					"format": "date-time"
+				},
+				"runQualities": {
+					"description": "List of runQuality elements.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/runQuality"
+					}
+				},
+				"setQualities": {
+					"description": "List of setQuality elements.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/setQuality"
+					}
+				},
+				"controlledVocabularies": {
+					"description": "Collection of CV elements.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/controlledVocabulary"
+					}
+				}
+			},
+			"additionalProperties": false,
+			"anyOf": [
+				{"required": ["runQualities"]},
+				{"required": ["setQualities"]}
+			],
+			"required": ["version", "controlledVocabularies"]
+		}
+	},
+	"additionalProperties": false,
+	"required": ["mzQC"],
+	"definitions": {
+		"baseQuality": {
+			"description": "Base element describing a runQuality or setQuality.",
+			"type": "object",
+			"properties": {
+				"metadata": {
+					"$ref": "#/definitions/metadata"
+				},
+				"qualityMetrics": {
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/qualityMetric"
+					}
+				}
+			},
+			"additionalProperties": false,
+			"required": ["metadata", "qualityMetrics"]
+		},
+		"runQuality": {
+			"description": "Parent element describing QC metrics for a single run.",
+			"$ref": "#/definitions/baseQuality"
+		},
+		"setQuality": {
+			"description": "Parent element describing combined QC metrics for multiple runs.",
+			"$ref": "#/definitions/baseQuality"
+		},
+		"cvParameter": {
+			"description": "Base element containing a reference to a controlled vocabulary.",
+			"type": "object",
+			"properties": {
+				"cvRef": {
+					"description": "Reference to the CV that contains the parameter definition.",
+					"type": "string",
+					"pattern": "^[A-Z]+$"
+				},
+				"accession": {
+					"description": "Accession number identifying the parameter within its CV.",
+					"type": "string",
+					"pattern": "^[A-Z]+:[0-9]{7}$"
+				},
+				"name": {
+					"description": "Name of the CV element describing the parameter.",
+					"type": "string"
+				},
+				"description": {
+					"description": "Description of the parameter.",
+					"type": "string"
+				},
+				"value": {
+					"description": "Value of the parameter."
+				},
+				"unit": {
+					"description": "A CV element describing the unit of the parameter value.",
+					"anyOf": [
+						{
+							"$ref": "#/definitions/cvParameter"
+						},
+						{
+							"type": "array",
+							"minItems": 1,
+							"items": {
+								"$ref": "#/definitions/cvParameter"
+							}
+						}
+					]
+				}
+			},
+			"required": ["cvRef", "accession", "name"]
+		},
+		"metadata": {
+			"description": "Metadata describing the QC analysis.",
+			"type": "object",
+			"properties": {
+				"fileProvenance": {
+					"description": "Optional string describing/documenting the provenance of the file",
+					"type": "string"
+				},
+				"inputFiles": {
+					"description": "List of input files based on which the QC metrics were generated.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/inputFile"
+					}
+				},
+				"analysisSoftware": {
+					"description": "Software tool(s) used to generate the QC metrics.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"allOf": [
+							{
+								"$ref": "#/definitions/cvParameter"
+							},
+							{
+								"properties": {
+									"version": {
+										"description": "Version number of the software tool.",
+										"type": "string"
+									},
+									"uri": {
+										"description": "Publicly accessible URI of the software tool (documentation).",
+										"type": "string",
+										"format": "uri"
+									}
+								},
+								"required": ["version", "uri"]
+							}
+						]
+					}
+				},
+				"cvParameters": {
+					"description": "Optional cvParameter tags for the QC analysis.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/cvParameter"
+					}
+				}
+			},
+			"additionalProperties": false,
+			"required": ["inputFiles", "analysisSoftware"]
+		},
+		"inputFile": {
+			"description": "File based on which QC metrics were generated.",
+			"type": "object",
+			"properties": {
+				"location": {
+					"description": "Unique file location, preferably publicly accessible.",
+					"type": "string",
+					"format": "uri"
+				},
+				"name": {
+					"description": "Base file name, should be unique across the mzQC file.",
+					"type": "string"
+				},
+				"fileFormat": {
+					"description": "Type of input file.",
+					"$ref": "#/definitions/cvParameter"
+				},
+				"fileProperties": {
+					"description": "Detailed properties of the input file.",
+					"type": "array",
+					"minItems": 1,
+					"items": {
+						"$ref": "#/definitions/cvParameter"
+					}
+				}
+			},
+			"additionalProperties": false,
+			"required": ["location", "name", "fileFormat"]
+		},
+		"qualityMetric": {
+			"description": "Parameter describing a QC metric.",
+			"$ref": "#/definitions/cvParameter"
+		},
+		"controlledVocabulary": {
+			"description": "Element describing a controlled vocabulary.",
+			"type": "object",
+			"properties": {
+				"name": {
+					"description": "Full name of the CV.",
+					"type": "string"
+				},
+				"ref": {
+					"description": "Reference string to denote the use of this specific CV by.",
+					"type": "string"
+				},
+				"uri": {
+					"description": "Publicly accessible URI of the CV.",
+					"type": "string",
+					"format": "uri"
+				},
+				"version": {
+					"description": "Version of the CV.",
+					"type": "string"
+				}
+			},
+			"additionalProperties": false,
+			"required": ["name", "uri"]
+		}
+	}
+}
